accordion prototype is completed
