--- src/components/AccordionWithButton/index.tsx
@@ -1,3 +1,0 @@
-import AccordionWithButton from './styles';
-
-export default AccordionWithButton;
