+++ src/pages/cli/graphql/authorization-rules.mdx
@@ -48,10 +48,6 @@ import accordion from '/src/fragments/cli/graphql/authorization_strategies.mdx';
 
 <Fragments fragments={{ all: accordion }} />
 
-import accordionWithButton from '/src/fragments/cli/graphql/authorization_strategies_button.mdx';
-
-<Fragments fragments={{ all: accordionWithButton }} />
-
 ## Authorization strategies
 
 Use the guide below to select the correct authorization strategy for your use case:
